requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beanie==2.0.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.2
h11==0.16.0
idna==3.10
lazy-model==0.3.0
motor==3.7.1
pip==25.2
pydantic==2.11.9
pydantic_core==2.33.2
pymongo==4.15.1
python-dotenv==1.1.1
setuptools==78.1.1
sniffio==1.3.1
starlette==0.48.0
typing_extensions==4.15.0
typing-inspection==0.4.1
uvicorn==0.35.0
wheel==0.45.1